requirements.txt
# Core dependencies
Flask==3.1.0
flask-cors==5.0.1
python-dotenv==1.0.1
requests==2.32.3
Werkzeug==3.0.1

# Payment processing
stripe==11.5.0

# Rate limiting and caching
Flask-Limiter==3.5.0
redis>=5.0.0
limits>=3.9.0

# Email handling
secure-smtplib>=0.1.1
email-validator>=2.1.0

# Security
itsdangerous==2.1.2
cryptography==42.0.0

# Server deployment
gunicorn==21.2.0
asgiref==3.8.1

# Development and testing
pytest>=8.0.0
pytest-cov>=4.1.0
black>=24.1.0
flake8>=7.0.0
isort>=5.13.2

# Documentation
Sphinx>=7.2.6
sphinx-rtd-theme>=2.0.0
